# src/integration-test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: json-schema
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:51633
  description: Generated server url
paths:
  /{municipalityId}/schemas/{schemaId}/ui-schema:
    get:
      tags:
      - UI-schemas
      summary: Get the UI schema that belongs to the provided JSON schema
      operationId: getUiSchemaById
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          type: string
          minLength: 1
        example: 2281_person_1.0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiSchema"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    put:
      tags:
      - UI-schemas
      summary: Create or replace a UI schema for the provided JSON schema
      operationId: createOrUpdateUiSchema
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          type: string
          minLength: 1
        example: 2281_person_1.0
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UiSchemaRequest"
        required: true
      responses:
        "204":
          description: No content - Successful operation
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      tags:
      - UI-schemas
      summary: Delete the UI schema for the provided JSON schema
      operationId: deleteUiSchema
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: schemaId
        in: path
        description: Schema ID
        required: true
        schema:
          type: string
          minLength: 1
        example: 2281_person_1.0
      responses:
        "204":
          description: No content - Successful operation
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    JsonNode: {}
    UiSchemaRequest:
      type: object
      description: JsonSchemaCreateRequest model
      properties:
        value:
          $ref: "#/components/schemas/JsonNode"
          description: The UI schema
          example: |
            {
              "ui:title": "Title",
              "ui:description": "Description",
              "ui:submitButtonOptions": {
                "props": {
                  "disabled": false,
                  "className": "btn btn-info",
                },
                  "norender": false,
                  "submitText": "Submit"
                }
            }
        description:
          type: string
          description: Description of the UI schema purpose
          examples:
          - A UI-schema that defines the rendering of the person form
      required:
      - value
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause: {}
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    UiSchema:
      type: object
      description: UI schema model
      properties:
        id:
          type: string
          description: UI Schema ID
          examples:
          - f83529c1-2dff-4a5d-aab0-8d4ec082a995
          readOnly: true
        value:
          $ref: "#/components/schemas/JsonNode"
          description: The UI schema
          example: |
            {
              "ui:title": "Title",
              "ui:description": "Description",
              "ui:submitButtonOptions": {
                "props": {
                  "disabled": false,
                  "className": "btn btn-info",
                },
                  "norender": false,
                  "submitText": "Submit"
                }
            }
          readOnly: true
        description:
          type: string
          description: Description of the UI schema purpose
          examples:
          - An UI-schema that defines how the form in the web-app should be rendered
          readOnly: true
        created:
          type: string
          format: date-time
          description: Created timestamp
          readOnly: true
      readOnly: true
  securitySchemes: {}